+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.4.0-dev
+
+### Enhancements
+  * Added support for [GitHub](https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#alerts)
+    and [GitLab](https://docs.gitlab.com/user/markdown/#alerts) alerts.
+  * Process alerts by default in Sigils.
+  * Added `:experimental_minimize_commonmark` render option.
+
 ## 0.3.3 (2025-02-03)
 
 ### Fixes